Updated genmessages build for Windows
--HG--
extra : convert_revision : svn%3Aad96d65e-77c6-4fb4-b022-7d5d8cc5d1e0/trunk%40201

+++ app/help/genmessages.c
@@ -68,7 +68,7 @@ TranslateString( char *srcString, struct transTbl *trTbl )
 	char *destString;
    char *cp;
    char *cp2;
-   int bufLen = strlen( srcString ) + 1;
+   size_t bufLen = strlen( srcString ) + 1;
    char *idx;
 
    /* calculate the expected result length */
